feat(elreal): widen block exponent to integer<256> per McCleeary's unbounded-exponent design (#1066)
Carry the per-block exponent in integer<256, uint32_t> (a trivially-copyable POD, so block stays hardware-shareable per #925) instead of int32, following McCleeary's unbounded-exponent design. Unblocks multi-block streaming division (#1061). Review: 5 CodeRabbit comments on exponent narrowing addressed (wide-type where possible, fail-loud guards in verification/production paths). Full elreal suite 34/34 on gcc+clang; full+fast CI tiers green (long sanitizer/coverage/POWER gates allowed to finish post-merge). Relates to #1061.

+++ include/sw/universal/number/elreal/divide.hpp
@@ -93,7 +93,7 @@ inline ZBCL<FpType> div(ZBCL<FpType> x, ZBCL<FpType> y, std::size_t depth = 64) 
         if (rem.empty()) break;                  // exact division: remainder is 0
         const B r0 = rem.front();                // leading block (descending order)
         if (!r0.is_normalised() || r0.exponent() < exp_floor) break;
-        const int lead = r0.exponent();
+        const int lead = static_cast<int>(r0.exponent());
         if (havePrev && lead >= prevLead) break; // no progress -> stop refining
         prevLead = lead;
         havePrev = true;
